Update GetMemoryMapWithPopulatedAccessAttributes() to Defer Free/Allocate Calls
Description

During the initialization of the memory protection, a complete map
of system memory is created via a call toGetMemoryMapWithPopulatedAccessAttributes().
The first step of the function was to fetch the EFI memory map and subsequent
steps manipulated it to create the output. The subsequent steps included
freeing and allocating memory which potentially changed the memory
layout and caused the output memory map to be incorrect. Because
allocating and freeing memory only transitions memory type between
EfiConventionalMemory and EfiBootServicesMemory (which are required
to have the same memory protection policy), this nuance didn't matter.
However, as the DXE core is updated to support more strict protections
on free memory, this nuance will matter.

This patch updates the GetMemoryMapWithPopulatedAccessAttributes()
routine to calculate and allocate all required memory before
fetching the EFI memory map used in the output, and adds a function
to free that memory once the caller is done with the ouput.

- [x] Impacts functionality?
  - **Functionality** - Does the change ultimately impact how firmware functions?
  - Examples: Add a new library, publish a new PPI, update an algorithm, ...
- [ ] Impacts security?
  - **Security** - Does the change have a direct security impact on an application,
    flow, or firmware?
  - Examples: Crypto algorithm change, buffer overflow fix, parameter
    validation improvement, ...
- [ ] Breaking change?
  - **Breaking change** - Will anyone consuming this change experience a break
    in build or boot behavior?
  - Examples: Add a new library class, move a module to a different repo, call
    a function in a new library class in a pre-existing module, ...
- [ ] Includes tests?
  - **Tests** - Does the change include any explicit test code?
  - Examples: Unit tests, integration tests, robot tests, ...
- [ ] Includes documentation?
  - **Documentation** - Does the change contain explicit documentation additions
    outside direct code modifications (and comments)?
  - Examples: Update readme file, add feature readme file, link to documentation
    on an a separate Web page, ...

How This Was Tested

Run on Q35

Integration Instructions

N/A